Moved normalized_object_type, find_h5_file and find_aggregation_netcdf
