+++ views/result_selection.py
@@ -18,9 +18,9 @@ from qgis.PyQt import uic
 
 from ..datasource.netcdf import (
     find_id_mapping_file,
-    find_h5_file,
     detect_netcdf_version,
 )
+from datasource.netcdf_groundwater import find_h5_file
 from datasource.result_constants import layer_qh_type_mapping
 from ..utils.user_messages import pop_up_info
 from .log_in_dialog import LoginDialog
